Optimisation, 8/16 bit values -> 32 bit values.
The Esp8266 (lx106) is a pure 32-bit architecture. Any references to
values smaller than 32 bits must be emulated. This results in larger and
slower code. Don't optimise on every byte and use 32-bit values wherever
possible appears to save 1-2 kbytes of code.

+++ application-parameters.h
@@ -25,9 +25,9 @@ typedef char args_t[application_num_args][application_length_args];
 typedef struct
 {
 	const char *cmdline;
-	uint8_t nargs;
+	unsigned int nargs;
 	args_t *args;
-	uint16_t size;
+	unsigned int size;
 	char *dst;
 } application_parameters_t;
